query parameters and split into front-end/back-end
