+++ package.json
@@ -16,7 +16,7 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node menu.js"
   },
   "repository": {
     "type": "git",
